+++ src/reg_tree.rs
@@ -5,7 +5,7 @@ use std::{
     ops::{Add, Div},
 };
 
-use crate::{float64::Float64, node::Node};
+use crate::{float32::Float32, node::Node};
 use conv::ValueFrom;
 use core::hash::Hash;
 use ndarray::prelude::*;
@@ -354,14 +354,14 @@ where
     }
 }
 
-impl Div<Float64> for usize {
-    type Output = Float64;
-    fn div(self, rhs: Float64) -> Self::Output {
-        <usize as Into<Float64>>::into(self) / rhs
+impl Div<Float32> for usize {
+    type Output = Float32;
+    fn div(self, rhs: Float32) -> Self::Output {
+        <usize as Into<Float32>>::into(self) / rhs
     }
 }
 
-impl Into<usize> for Float64 {
+impl Into<usize> for Float32 {
     fn into(self) -> usize {
         self.0 as usize
     }
@@ -372,16 +372,16 @@ mod tests {
 
     use ndarray_npy::read_npy;
 
-    use crate::float64::Float64;
+    use crate::float32::Float32;
 
     use super::*;
 
     #[test]
     fn easiest_test() {
-        type G = Float64;
-        type R = Float64;
+        type G = Float32;
+        type R = Float32;
         let data = array![[1, 1, 1, 1], [2, 2, 2, 2], [3, 3, 3, 3], [4, 4, 4, 4]];
-        let data = data.mapv(|x| Float64(x as f32));
+        let data = data.mapv(|x| Float32(x as f32));
         let target = array![-1, -1, 1, 1];
 
         let mut tree: RegressionTree<G, R> = RegressionTree {
@@ -396,12 +396,12 @@ mod tests {
 
     #[test]
     fn test() {
-        type G = Float64;
-        type R = Float64;
+        type G = Float32;
+        type R = Float32;
 
-        let data: Array2<f32> =
+        let data: Array2<f64> =
             read_npy("src/three_9_data.npy").expect("file should be present and correct");
-        let data = data.mapv(|x| Float64(x as f32));
+        let data = data.mapv(|x| Float32(x as f32));
 
         let target: Array1<i64> =
             read_npy("src/three_9_target.npy").expect("file should be present and correct");
